Refactor PreferenceDialog to MVVM with a split-pane
