+++ src/main/java/org/jabref/gui/preferences/PreferencesDialog.fxml
@@ -1,5 +1,52 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
-
+<?import javafx.geometry.Insets?>
+<?import javafx.scene.control.Button?>
+<?import javafx.scene.control.ButtonType?>
 <?import javafx.scene.control.DialogPane?>
-<DialogPane prefHeight="700.0" prefWidth="1100.0" xmlns="http://javafx.com/javafx/8.0.171" xmlns:fx="http://javafx.com/fxml/1" fx:controller="org.jabref.gui.preferences.PreferencesDialog" />
+<?import javafx.scene.control.ListView?>
+<?import javafx.scene.control.ScrollPane?>
+<?import javafx.scene.control.SplitPane?>
+<?import javafx.scene.control.TextField?>
+<?import javafx.scene.control.Tooltip?>
+<?import javafx.scene.layout.VBox?>
+
+<DialogPane prefHeight="700.0" prefWidth="1100.0" minHeight="400" minWidth="600" xmlns="http://javafx.com/javafx/8.0.212" xmlns:fx="http://javafx.com/fxml/1" fx:controller="org.jabref.gui.preferences.PreferencesDialogView">
+    <content>
+        <SplitPane dividerPositions="0.3">
+            <items>
+                <VBox prefWidth="160.0" styleClass="split-pane">
+                    <children>
+                        <TextField fx:id="searchBox" promptText="%Search" VBox.vgrow="NEVER">
+                            <VBox.margin>
+                                <Insets bottom="3.0" left="3.0" right="3.0" top="3.0" />
+                            </VBox.margin>
+                        </TextField>
+                        <ListView fx:id="preferenceTabList" VBox.vgrow="ALWAYS" />
+                        <VBox prefHeight="10.0" VBox.vgrow="SOMETIMES" />
+                        <VBox alignment="BOTTOM_LEFT" spacing="3.0">
+                            <children>
+                                <Button maxWidth="Infinity" onAction="#importPreferences" text="%Import preferences">
+                                    <tooltip><Tooltip text="%Import preferences from file" /></tooltip>
+                                </Button>
+                                <Button maxWidth="Infinity" onAction="#exportPreferences" text="%Export preferences">
+                                    <tooltip><Tooltip text="%Export preferences to file" /></tooltip>
+                                </Button>
+                                <Button maxWidth="Infinity" onAction="#showAllPreferences" text="%Show preferences" />
+                                <Button maxWidth="Infinity" onAction="#resetPreferences" text="%Reset preferences" />
+                            </children>
+                            <padding>
+                                <Insets bottom="3.0" left="3.0" right="3.0" top="3.0" />
+                            </padding>
+                        </VBox>
+                    </children>
+                </VBox>
+                <ScrollPane fx:id="preferencePaneContainer" />
+            </items>
+        </SplitPane>
+    </content>
+    <buttonTypes>
+        <ButtonType fx:id="saveButton" text="%Save" buttonData="OK_DONE" />
+        <ButtonType fx:constant="CANCEL" />
+    </buttonTypes>
+</DialogPane>
