+++ src/main/java/org/jabref/gui/preferences/PreferencesDialogView.java
@@ -1,0 +1,132 @@
+package org.jabref.gui.preferences;
+
+import java.util.Locale;
+
+import javax.inject.Inject;
+
+import javafx.fxml.FXML;
+import javafx.scene.control.ButtonType;
+import javafx.scene.control.ListView;
+import javafx.scene.control.ScrollPane;
+import javafx.scene.control.TextField;
+
+import org.jabref.gui.DialogService;
+import org.jabref.gui.JabRefFrame;
+import org.jabref.gui.util.BaseDialog;
+import org.jabref.gui.util.ControlHelper;
+import org.jabref.gui.util.TaskExecutor;
+import org.jabref.gui.util.ViewModelListCellFactory;
+import org.jabref.logic.l10n.Localization;
+import org.jabref.preferences.PreferencesService;
+
+import com.airhacks.afterburner.views.ViewLoader;
+import org.fxmisc.easybind.EasyBind;
+
+/**
+ * Preferences dialog. Contains a TabbedPane, and tabs will be defined in separate classes. Tabs MUST implement the
+ * PrefsTab interface, since this dialog will call the storeSettings() method of all tabs when the user presses ok.
+ */
+public class PreferencesDialogView extends BaseDialog<PreferencesDialogViewModel> {
+
+    @FXML private TextField searchBox;
+    @FXML private ListView<PrefsTab> preferenceTabList;
+    @FXML private ScrollPane preferencePaneContainer;
+    @FXML private ButtonType saveButton;
+
+    @Inject private DialogService dialogService;
+    @Inject private PreferencesService preferences;
+
+    private JabRefFrame frame;
+    private TaskExecutor taskExecutor;
+    private PreferencesDialogViewModel viewModel;
+
+    public PreferencesDialogView(JabRefFrame frame, TaskExecutor taskExecutor) {
+        this.frame = frame;
+        this.taskExecutor = taskExecutor;
+        this.setTitle(Localization.lang("JabRef preferences"));
+
+        ViewLoader.view(this)
+                  .load()
+                  .setAsDialogPane(this);
+
+        ControlHelper.setAction(saveButton, getDialogPane(), event -> savePreferencesAndCloseDialog());
+    }
+
+    public PreferencesDialogViewModel getViewModel() { return viewModel; }
+
+    @FXML
+    private void initialize () {
+        viewModel = new PreferencesDialogViewModel(dialogService, taskExecutor, frame, this);
+        getDialogPane().getScene().getStylesheets().add(this.getClass().getResource("PreferencesDialog.css").toExternalForm());
+        preferencePaneContainer.getStyleClass().add("preferencePaneContainer");
+
+        preferenceTabList.itemsProperty().setValue(viewModel.getPreferenceTabs());
+
+        PreferencesSearchHandler searchHandler = new PreferencesSearchHandler(viewModel.getPreferenceTabs());
+        preferenceTabList.itemsProperty().bindBidirectional(searchHandler.filteredPreferenceTabsProperty());
+        searchBox.textProperty().addListener((observable, previousText, newText) -> {
+            searchHandler.filterTabs(newText.toLowerCase(Locale.ROOT));
+            preferenceTabList.getSelectionModel().clearSelection();
+            preferenceTabList.getSelectionModel().selectFirst();
+        });
+
+        EasyBind.subscribe(preferenceTabList.getSelectionModel().selectedItemProperty(), tab -> {
+            if (tab != null) {
+                preferencePaneContainer.setContent(tab.getBuilder());
+            } else {
+                preferencePaneContainer.setContent(null);
+            }
+        });
+
+        preferenceTabList.getSelectionModel().selectFirst();
+        new ViewModelListCellFactory<PrefsTab>()
+                .withText(PrefsTab::getTabName)
+                .install(preferenceTabList);
+
+        setValues();
+    }
+
+    @FXML
+    private void closeDialog() { close(); }
+
+    @FXML
+    private void savePreferencesAndCloseDialog() {
+        viewModel.storeAllSettings();
+        closeDialog();
+    }
+
+    @FXML
+    void exportPreferences() {
+        viewModel.exportPreferences();
+    }
+
+    @FXML
+    void importPreferences() {
+        viewModel.importPreferences();
+    }
+
+    @FXML
+    void showAllPreferences() {
+        viewModel.showPreferences();
+    }
+
+    @FXML
+    void resetPreferences() {
+        viewModel.resetPreferences();
+    }
+
+    /**
+     * Sets the values in each preference tab according to Globals.prefs.
+     * Currently the PreferenceDialogViewModel uses this method, which,
+     * I believe, is not allowed in the MVVM-model. Fixing this would
+     * probably require refactoring the JabRefPreferences to use
+     * Properties probably (?), which would go way beyond this PR.
+     * So this is just deprecated and awaits refactoring.
+     */
+    @Deprecated
+    public void setValues() {
+        for (PrefsTab prefsTab : viewModel.getPreferenceTabs()) {
+            prefsTab.setValues();
+        }
+    }
+}
